+++ src/components/login/Login.jsx
@@ -1,6 +1,6 @@
 import React, {useState, useEffect} from 'react'
 import Button from 'react-bootstrap/Button'
-import { useModal } from '../hooks/useModal';
+import { useModal } from '../../hooks/useModal';
 import Form from 'react-bootstrap/Form';
 import './login.css'
 import instance from '../../api'
